# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/AdminListings": {
            "get": {
                "tags": [
                    "Admin Route"
                ],
                "summary": "Summaries",
                "description": "AdminDescription",
                "operationId": "Titlings",
                "responses": {
                    "200": {
                        "description": "Successfully retreived the data"
                    }
                }
            }
        },
        "/api/RegisteringAdmin": {
            "post": {
                "tags": [
                    "Admin Route"
                ],
                "summary": "Summaries",
                "description": "AdminDescriptions",
                "operationId": "Tittlin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Message",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/userList": {
            "get": {
                "tags": [
                    "User Route"
                ],
                "summary": "registering a new user",
                "description": "Description",
                "operationId": "TestId",
                "responses": {
                    "200": {
                        "description": "Successfully retreived the data"
                    }
                }
            }
        },
        "/api/registerUser": {
            "post": {
                "tags": [
                    "User Route"
                ],
                "summary": "User Registration",
                "description": "Description",
                "operationId": "Title",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Message",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/updateUser/{id}": {
            "put": {
                "tags": [
                    "User Route"
                ],
                "summary": "Summary",
                "description": "Description",
                "operationId": "Titles",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Message",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}